feat: 네트워크 연결 끊김 화면 메소드 생성 #60
